+++ crates/bif_qt/cpp/property_inspector_widget.h
@@ -18,7 +18,6 @@
 #include <QWidget>
 
 class BifShellState;
-class QCheckBox;
 class QComboBox;
 class QGroupBox;
 class QLabel;
@@ -39,7 +38,6 @@ public:
 private slots:
     void on_selection_changed();
     void on_layer_state_changed();
-    void on_visibility_toggled(bool checked);
     void on_bind_material_clicked();
     void on_shading_model_changed(const QString& model);
 
@@ -53,7 +51,6 @@ private:
 
     QLabel* m_header_path;
     QLabel* m_header_type;
-    QCheckBox* m_visibility_box;
     QGroupBox* m_arcs_group;
     QListWidget* m_arcs_list;
     QTabWidget* m_tabs;