+++ crates/bif_qt/cpp/scene_browser_model.h
@@ -36,7 +36,8 @@ public:
     };
 
     enum Columns {
-        ColName = 0,
+        ColVisibility = 0,
+        ColName,
         ColType,
         ColChildren,
         ColKind,